pyproject.toml
[project]
name = "ai-perso-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "faiss-cpu",
    "python-dotenv",
    "openai",
    "sentence-transformers",
]
